+++ src/main/java/bsise/server/clova/dto/ClovaRequestDto.java
@@ -1,4 +1,4 @@
-package bsise.server.clovar;
+package bsise.server.clova.dto;
 
 public interface ClovaRequestDto {
 }
