+++ src/main/java/bsise/server/clova/service/ClovaService.java
@@ -1,13 +1,20 @@
-package bsise.server.clovar;
+package bsise.server.clova.service;
 
 import static bsise.server.user.domain.Preference.*;
 
-import bsise.server.clovar.dailyReport.ClovaDailyReportRequestDto;
+import bsise.server.clova.dto.ClovaLetterReplyRequestDto;
+import bsise.server.clova.dto.ClovaRequestDto;
+import bsise.server.clova.dto.ClovaResponseDto;
+import bsise.server.clova.dto.TwoTypeMessage;
+import bsise.server.clova.client.ClovaFeignClient;
+import bsise.server.clova.dailyReport.ClovaDailyReportRequestDto;
 import bsise.server.report.weekly.dto.ClovaWeeklyReportRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
+@Profile("prod")
 @Service
 @RequiredArgsConstructor
 public class ClovaService {
