+++ src/main/java/bsise/server/clova/service/DummyReportClavaService.java
@@ -1,6 +1,8 @@
-package bsise.server.clovar;
+package bsise.server.clova.service;
 
-import bsise.server.clovar.dailyReport.DummyDailyReportClovaResponseDto;
+import bsise.server.clova.dto.ClovaResponseDto;
+import bsise.server.clova.client.ClovaFeignClient;
+import bsise.server.clova.dailyReport.DummyDailyReportClovaResponseDto;
 import bsise.server.report.weekly.dto.ClovaWeeklyReportRequestDto;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
